+++ src/app/admin-tools/managers-list/manager-options/manager-options.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {EmployeesService} from '../../../employees.service';
 import {ActivatedRoute, Params, Router} from '@angular/router';
 import {UtilService} from '../../../util.service';
@@ -11,9 +11,11 @@ import {EmployeeDto} from '../../../models-container.model';
 })
 export class ManagerOptionsComponent implements OnInit {
   manager: EmployeeDto;
+  error: string;
 
   constructor(private employeesService: EmployeesService,
-              private route: ActivatedRoute, private router: Router) {}
+              private route: ActivatedRoute, private router: Router) {
+  }
 
   ngOnInit() {
     this.route.params.subscribe(
@@ -25,18 +27,25 @@ export class ManagerOptionsComponent implements OnInit {
   }
 
   onRemove() {
-    this.employeesService.removeEmployee(this.manager.id)
-      .subscribe(
-        () => UtilService.redirectTo('/admin-tools/managers-list', this.router),
-        error => console.log(error)
-      );
+    if (confirm('ARE YOU SURE TO DELETE ' + this.manager.firstName
+      + ' ' + this.manager.lastName + ' FROM BASE?')) {
+      this.employeesService.removeEmployee(this.manager.id)
+        .subscribe(
+          () => UtilService.redirectTo('/admin-tools/managers-list', this.router),
+          error => this.error = error
+        );
+    }
   }
 
   onChangeRole() {
-    this.employeesService.changeRole(this.manager.id)
-      .subscribe(
-        () => UtilService.redirectTo('/admin-tools/managers-list', this.router),
-        error => console.log(error)
-      );
+    if (confirm('ARE YOU SURE TO MAKE ' + this.manager.firstName
+      + ' ' + this.manager.lastName + ' SPECIALIST ? DATA ABOUT SUBORDINATES WILL BE LOST!')) {
+      this.employeesService.changeRole(this.manager.id)
+        .subscribe(
+          () => UtilService.redirectTo('/admin-tools/managers-list', this.router),
+          error => this.error = error
+        );
+    }
   }
+
 }
